simplified: module rearrangement
