--- crates/producers/disolv-positions/src/activation/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod cache;
-pub(crate) mod writer;
